Crate wit_writer

A crate to write the raw wasm interface types section.

Structs

Exports
Writer for the list of exports in an export subsection.

Funcs
Writer for the list of functions in an function subsection.

Implements
Writer for the list of imports in an import subsection.

Imports
Writer for the list of imports in an import subsection.

Instructions
Writer for a sequence of instructions in a function subsection

Type
A writer to write out a type, or a sequence of types.

Types
Writer for the list of types in a type subsection.

Writer
A structure used to write out the raw representation of a wasm interface types subsection.